Serialization now works (hopefully)
I was expecting strings to have some corner cases because of null
characters, and iterators just work. After 12 hours of debugging, and
then a bugreport to Cereal, and a discussion, turned out it's upside
down: iterators randomly screw binary data, but std::string work with it
just fine.

Also do some constification while on it.

Signed-off-by: Konstantin Kharlamov <Hi-Angel@yandex.ru>