Fix build errors after merge
The cause for them is unknown, the preprocessor upon building produces
for example the following three lines:

    std::pair<int,std::vector<uint8_t>> inflate(const std::vector<uint8_t> in);
    # 366 "../protocol.h" 2
    std::pair<int,std::vector<uint8_t>> inflate(const std::vector<uint8_t> in);

Upon removing the second declaration gcc starts failing to see the first
one. This is almost definitely a bug in GCC, unfortunately I am too low
on time for reducing the code to minimal example for reporting, so let's
just make it working.

Signed-off-by: Konstantin Kharlamov <Hi-Angel@yandex.ru>